refactor(docs): extract theme configuration into separate file and update vitepress config
